% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcflightsql-package.R
\name{adbcflightsql}
\alias{adbcflightsql}
\alias{adbc_database_init.adbcflightsql_driver_flightsql}
\alias{adbc_connection_init.adbcflightsql_database}
\alias{adbc_statement_init.adbcflightsql_connection}
\title{ADBC FlightSQL Driver}
\usage{
adbcflightsql()

\method{adbc_database_init}{adbcflightsql_driver_flightsql}(driver, ..., uri = NULL)

\method{adbc_connection_init}{adbcflightsql_database}(database, ..., adbc.connection.autocommit = NULL)

\method{adbc_statement_init}{adbcflightsql_connection}(
  connection,
  ...,
  adbc.ingest.target_table = NULL,
  adbc.ingest.mode = NULL
)
}
\arguments{
\item{driver}{The driver to use. This can be one of the following:
\itemize{
\item A non-missing \code{character(1)} containing a driver or manifest name, a
relative or absolute path to a driver or manifest, or a URI. For a URI,
the driver manager uses the URI scheme as the driver name and passes the
URI to that driver. A \verb{profile://} URI loads a connection profile.
\item \code{NULL}, which leaves driver selection to the driver manager. In this
case, \code{...} must normally contain \code{uri} or \code{profile}; the driver is
inferred from the URI or loaded from the connection profile.
\item An object that inherits from the \code{adbc_driver} class, such as one created
by \code{\link[adbcdrivermanager:adbc_driver]{adbc_driver()}}. This includes drivers provided by R packages, such as
\code{adbcsqlite::adbcsqlite()}.
}}

\item{...}{Extra key/value options passed to the driver.}

\item{uri}{A URI to a database path (e.g.,
\verb{user[:password]@account/database[?param1=value1]})}

\item{database}{An \link[adbcdrivermanager:adbc_database_init]{adbc_database}.}

\item{adbc.connection.autocommit}{Use FALSE to disable the default
autocommit behaviour.}

\item{connection}{An \link[adbcdrivermanager:adbc_connection_init]{adbc_connection}}

\item{adbc.ingest.target_table}{The name of the target table for a bulk insert.}

\item{adbc.ingest.mode}{Whether to create (the default) or append.}
}
\value{
An \code{\link[adbcdrivermanager:adbc_driver]{adbcdrivermanager::adbc_driver()}}
}
\description{
ADBC FlightSQL Driver
}
\examples{
adbcflightsql()

}
